feat: add product search (EA-103)
