+++ src/app/pages/catalog/catalog-page.ts
@@ -1,43 +1,35 @@
 import { Page } from '@templates/page';
 import { NotFoundPage } from '@servicePages/404-page';
-import { el, mount, unmount, setChildren } from 'redom';
-import {
-  getProducts,
-  getCategories,
-  getProductsOfCategory,
-  getCategoryByKey,
-  getFilteredProducts,
-} from '@sdk/requests';
+import { el, mount, setChildren } from 'redom';
+import { getProducts, getCategories, getProductsOfCategory, getCategoryByKey } from '@sdk/requests';
 import { ProductProjection } from '@commercetools/platform-sdk';
 import { ProductMainData } from '@customTypes/catalog';
-import { Colors } from '@customTypes/enums';
 import { ProductCard } from '@components/product-card';
-import { Button } from '@components/button';
+import { FiltersBlock } from '@components/filters-block';
 import { extractProductData, buildCategoriesObject } from '@helpers/catalog';
 import { safeQuerySelector } from '@helpers/safe-query-selector';
 import { router } from '@app/router';
+import magnifier from '@icons/magnifying-glass.svg';
 
 class CatalogPage extends Page {
   constructor(public categoryKey?: string) {
     super();
   }
 
+  public sideBar = el('.catalog-sidebar');
+  public searchInput = el('input.catalog-search-input', {
+    placeholder: 'Search...',
+  });
+
   protected textObject = {
     title: 'Catalog',
   };
 
   private productsContainer = el('.products');
-  private sideBar = el('.catalog-sidebar');
   private mask = el('.catalog-mask');
-  private minPrice = 0;
-  private maxPrice = 5;
-  private priceFrom = this.minPrice;
-  private priceTo = this.maxPrice;
-  private filtersPriceBlock = el('.filters-unit');
-  private filtersColorBlock = el('.filters-unit');
-  private colorCircles: HTMLElement[] = [];
+  private filtersBlock = new FiltersBlock(this);
 
-  protected createProductContainer(): HTMLElement {
+  public createProductContainer(): HTMLElement {
     this.getProductData().then((products) => {
       if (!products) {
         new NotFoundPage().render();
@@ -45,6 +37,49 @@ class CatalogPage extends Page {
       this.fillProductsContainer(products);
     });
     return this.productsContainer;
+  }
+
+  public listenTitleClick(title: HTMLElement): void {
+    const windowSize = window.matchMedia('(max-width: 768px)');
+    const mask = this.mask;
+    let isAllTitlesInactive: boolean;
+
+    title.addEventListener('click', () => {
+      title.classList.toggle('active');
+      const titles = document.querySelectorAll('.sidebar-dropdown-title');
+
+      titles.forEach((item) => {
+        if (item !== title) {
+          item.classList.remove('active');
+        } else if (item.classList.contains('active')) {
+          isAllTitlesInactive = false;
+        } else {
+          isAllTitlesInactive = true;
+        }
+      });
+      if (windowSize.matches) {
+        if (isAllTitlesInactive) {
+          mask.classList.remove('lock');
+          document.body.classList.remove('no-scroll');
+        } else {
+          mask.classList.add('lock');
+          document.body.classList.add('no-scroll');
+        }
+      }
+    });
+    windowSize.addEventListener('change', () => {
+      this.closeCategoryNav(title, mask);
+    });
+    this.closeCategoryNavOnBurgerClick(title, mask);
+  }
+
+  public fillProductsContainer(products: ProductMainData[]): void {
+    this.productsContainer.innerHTML = '';
+    products.forEach((product) => {
+      const card = new ProductCard(product).create();
+      mount(this.productsContainer, card);
+    });
+    router.updatePageLinks();
   }
 
   private async getProductData(): Promise<ProductMainData[]> {
@@ -90,7 +125,31 @@ class CatalogPage extends Page {
     return container;
   }
 
-  protected createCategoryList(): HTMLElement {
+  private createSearch(): HTMLElement {
+    const { searchInput } = this;
+
+    searchInput.addEventListener('keydown', (e) => {
+      if (e.key === 'Enter' && searchInput instanceof HTMLInputElement) {
+        const searchQuery = searchInput.value
+          .trim()
+          .toLowerCase()
+          .split(' ')
+          .map((userQuery) => `"${userQuery}"`)
+          .join(',');
+
+        this.filtersBlock.applyFilters(searchQuery);
+      }
+    });
+
+    return el('.catalog-search', [
+      searchInput,
+      el('img.catalog-search-icon', {
+        src: magnifier,
+      }),
+    ]);
+  }
+
+  private createCategoryList(): HTMLElement {
     const title = el('h3.categories-title.sidebar-dropdown-title', [el('span', 'Categories')]);
     const container = el('.categories', title);
     const categoriesList = el('ul.sidebar-dropdown-content');
@@ -132,40 +191,6 @@ class CatalogPage extends Page {
     return container;
   }
 
-  private listenTitleClick(title: HTMLElement): void {
-    const windowSize = window.matchMedia('(max-width: 768px)');
-    const mask = this.mask;
-    let isAllTitlesInactive: boolean;
-
-    title.addEventListener('click', () => {
-      title.classList.toggle('active');
-      const titles = document.querySelectorAll('.sidebar-dropdown-title');
-
-      titles.forEach((item) => {
-        if (item !== title) {
-          item.classList.remove('active');
-        } else if (item.classList.contains('active')) {
-          isAllTitlesInactive = false;
-        } else {
-          isAllTitlesInactive = true;
-        }
-      });
-      if (windowSize.matches) {
-        if (isAllTitlesInactive) {
-          mask.classList.remove('lock');
-          document.body.classList.remove('no-scroll');
-        } else {
-          mask.classList.add('lock');
-          document.body.classList.add('no-scroll');
-        }
-      }
-    });
-    windowSize.addEventListener('change', () => {
-      this.closeCategoryNav(title, mask);
-    });
-    this.closeCategoryNavOnBurgerClick(title, mask);
-  }
-
   private closeCategoryNav(title: HTMLElement, mask: HTMLElement): void {
     title.classList.remove('active');
     mask.classList.remove('lock');
@@ -186,145 +211,16 @@ class CatalogPage extends Page {
     });
   }
 
-  private createFiltersBlock(): HTMLElement {
-    const title = el('h3.filters-title.sidebar-dropdown-title', [el('span', 'Filters')]);
-    const applyBtn = new Button('Apply', 'black').render();
-    const resetBtn = new Button('Reset', 'black').render();
-    const priceFilters = this.createPriceRange();
-    const colorFilters = this.createColorFiltersBlock();
-
-    priceFilters?.classList.remove('applied');
-    colorFilters.classList.remove('applied');
-
-    const container = el('.filters', [
-      title,
-      el('.sidebar-dropdown-content', [priceFilters, colorFilters, el('.filters-buttons', [resetBtn, applyBtn])]),
-    ]);
-    applyBtn.addEventListener('click', () => {
-      this.applyFilters();
-    });
-    resetBtn.addEventListener('click', () => {
-      console.log('click');
-      if (!this.isPriceFiltersApplied() && !this.isColorFiltersApplied()) {
-        return;
-      }
-      unmount(this.sideBar, container);
-      mount(this.sideBar, this.createFiltersBlock());
-      this.createProductContainer();
-    });
-    this.listenTitleClick(title);
-    router.updatePageLinks();
-    return container;
-  }
-
-  private async applyFilters(): Promise<void> {
-    const queryArgs = this.assembleQueryArgs();
-    if (!queryArgs.length) {
-      return;
-    }
-    if (this.categoryKey) {
-      const category = await getCategoryByKey(this.categoryKey);
-      queryArgs.push(`categories.id:"${category?.id}"`);
-    }
-    const products = await extractProductData(getFilteredProducts(queryArgs));
-    this.fillProductsContainer(products);
-  }
-
-  private fillProductsContainer(products: ProductMainData[]): void {
-    this.productsContainer.innerHTML = '';
-    products.forEach((product) => {
-      const card = new ProductCard(product).create();
-      mount(this.productsContainer, card);
-    });
-    router.updatePageLinks();
-  }
-
-  private assembleQueryArgs(): string[] {
-    const queryArgs: string[] = [];
-
-    if (this.isPriceFiltersApplied()) {
-      this.filtersPriceBlock.classList.add('applied');
-      queryArgs.push(`variants.price.centAmount:range (${this.priceFrom * 100} to ${this.priceTo * 100})`);
-    } else {
-      this.filtersPriceBlock.classList.remove('applied');
-    }
-    if (this.isColorFiltersApplied()) {
-      const selectedColors = this.colorCircles
-        .filter((circle) => circle.classList.contains('active'))
-        .map((circle) => `"${circle.getAttribute('data-color')}"`)
-        .join(',');
-
-      this.filtersColorBlock.classList.add('applied');
-      queryArgs.push(`variants.attributes.color.key:${selectedColors}`);
-    } else {
-      this.filtersColorBlock.classList.remove('applied');
-    }
-    return queryArgs;
-  }
-
-  private isPriceFiltersApplied(): boolean {
-    return this.priceFrom !== this.minPrice || this.priceTo !== this.maxPrice;
-  }
-
-  private isColorFiltersApplied(): boolean {
-    return this.colorCircles.some((circle) => circle.classList.contains('active'));
-  }
-
-  private createPriceRange(): HTMLElement | void {
-    const thumbFrom = el('input#from', { type: 'range', min: this.minPrice, value: '0', max: this.maxPrice });
-    const thumbTo = el('input#to', { type: 'range', min: this.minPrice, value: '5', max: this.maxPrice });
-
-    if (!(thumbFrom instanceof HTMLInputElement) || !(thumbTo instanceof HTMLInputElement)) {
-      return;
-    }
-
-    [thumbFrom, thumbTo].forEach((thumb) => {
-      thumb.addEventListener('input', () => {
-        const priceWrap = thumb.parentElement;
-        if (!(thumb instanceof HTMLInputElement)) {
-          return;
-        }
-        priceWrap?.style.setProperty(`--${thumb.id}`, thumb.value);
-
-        this.priceFrom = Math.min(+thumbFrom.value, +thumbTo.value);
-        this.priceTo = Math.max(+thumbFrom.value, +thumbTo.value);
-      });
-    });
-    setChildren(this.filtersPriceBlock, [
-      el('h4.filters-unit-title', 'Price'),
-      el('.filters-price-wrap', [
-        thumbFrom,
-        el('output.price-output-from', { for: 'from' }),
-        thumbTo,
-        el('output.price-output-to', { for: 'to' }),
-      ]),
-    ]);
-    return this.filtersPriceBlock;
-  }
-
-  private createColorFiltersBlock(): HTMLElement {
-    this.colorCircles = [];
-    for (const [color, hex] of Object.entries(Colors)) {
-      const circle = el('.filters-color-circle', { style: `background-color: ${hex}`, 'data-color': color });
-      circle.addEventListener('click', () => {
-        circle.classList.toggle('active');
-      });
-      this.colorCircles.push(circle);
-    }
-    setChildren(this.filtersColorBlock, [
-      el('h4.filters-colors-title.filters-unit-title', 'Colors'),
-      el('.filters-colors-content', this.colorCircles),
-    ]);
-    return this.filtersColorBlock;
-  }
-
   protected build(): HTMLElement {
-    setChildren(this.sideBar, [this.createCategoryList(), this.createFiltersBlock()]);
+    setChildren(this.sideBar, [this.createCategoryList(), this.filtersBlock.create()]);
 
     return el('section.catalog', [
       this.mask,
       this.sideBar,
-      el('.products-wrapper', [this.createBreadcrumbs(), this.createProductContainer()]),
+      el('.products-wrapper', [
+        el('.breadcrumbs-search-wrap', [this.createBreadcrumbs(), this.createSearch()]),
+        this.createProductContainer(),
+      ]),
     ]);
   }
 }
